vcl: support for unidirectional streams
Type: improvement

+++ src/vnet/session/application_interface.h
@@ -358,7 +358,7 @@ typedef struct session_accepted_msg_
   u32 mq_index;
   transport_endpoint_t lcl;
   transport_endpoint_t rmt;
-  u8 flags;
+  u16 flags;
   /* TODO(fcoras) maybe refactor to pass as transport attr */
   u32 original_dst_ip4;
   u16 original_dst_port;
